requirements.txt
Django==2.0.5
django-tastypie==0.14.1
djangorestframework==3.8.2
mysqlclient==1.3.12
python-dateutil==2.7.3
python-mimeparse==1.6.0
pytz==2018.4
six==1.11.0
